+++ Tests/TechTalk.SpecFlow.RuntimeTests/Bindings/AsyncHelperTests.cs
@@ -1,21 +1,27 @@
 ﻿using System;
+using System.Threading.Tasks;
 using FluentAssertions;
+using TechTalk.SpecFlow.Bindings;
 using Xunit;
 
 namespace TechTalk.SpecFlow.RuntimeTests.Bindings
 {
     public class AsyncHelperTests
     {
-        public void NormalMethod()
+        private void NormalMethod()
         {
 
         }
 
-        public void ThrowException()
+        private void ThrowException()
         {
             throw new Exception("Hi from async");
         }
 
+        private Task AsyncMethodThrowsException()
+        {
+            throw new Exception();
+        }
 
         [Fact]
         public void RunSync_NormalMethod()
@@ -30,6 +36,19 @@ namespace TechTalk.SpecFlow.RuntimeTests.Bindings
             Action action = () => AsyncHelpers.RunSync(async () => ThrowException());
 
             action.Should().Throw<Exception>().WithMessage("Hi from async");
+        }
+
+        [Fact]
+        public void RunSync_WhenExceptionIsThrown_ShouldRestoreOriginalStackTrace()
+        {
+            try
+            {
+                AsyncHelpers.RunSync(AsyncMethodThrowsException);
+            }
+            catch (Exception ex)
+            {
+                ex.StackTrace.Should().StartWith($"   at {GetType().FullName}.{nameof(AsyncMethodThrowsException)}()");
+            }
         }
     }
 }
